requirements.txt
langchain-openai
langgraph
playwright
pydantic
python-dotenv
streamlit
beautifulsoup4
Pillow
numpy
